+++ packages/app/src/services/thumbnail/thumbnailService.ts
@@ -1,27 +1,34 @@
-import { app } from 'electron';
 import { join } from 'node:path';
 import { Readable } from 'node:stream';
+import { app, dialog } from 'electron';
 import { USER_AGENT } from '~/constants';
-import { mkdir } from 'node:fs/promises';
+import { IpcService } from '~/services/ipc';
 import { createWriteStream } from 'node:fs';
 import { HttpService } from '~/services/http';
 import { finished } from 'node:stream/promises';
 import { LoggingService } from '~/services/logging';
 import { inject, injectable } from '@needle-di/core';
+import { mkdir, unlink, readdir } from 'node:fs/promises';
 import { dirExists, fileExists } from '@depthbomb/node-common/fs';
+import type { IBootstrappable } from '~/common';
 import type { HttpClient } from '~/services/http';
 
 @injectable()
-export class ThumbnailService {
+export class ThumbnailService implements IBootstrappable {
 	private readonly httpClient: HttpClient;
 	private readonly cacheDir: string;
 
 	public constructor(
+		private readonly ipc    = inject(IpcService),
 		private readonly logger = inject(LoggingService),
 		private readonly http   = inject(HttpService),
 	) {
 		this.httpClient = this.http.getClient('ThumbnailService', { userAgent: USER_AGENT });
 		this.cacheDir   = join(app.getPath('userData'), 'thumbnail_cache');
+	}
+
+	public async bootstrap() {
+		this.ipc.registerHandler('thumbnail<-clear-cache', () => this.clearCache());
 	}
 
 	public async downloadThumbnail(videoId: string) {
@@ -56,5 +63,27 @@ export class ThumbnailService {
 		}
 
 		return thumbnailPath;
+	}
+
+	private async clearCache() {
+		const files = await readdir(this.cacheDir);
+		if (!files.length) {
+			await dialog.showMessageBox({
+				title: 'Clear thumbnail cache',
+				message: 'No thumbnails to clear.'
+			});
+			return;
+		}
+
+		for (const file of files) {
+			await unlink(
+				join(this.cacheDir, file)
+			);
+		}
+
+		await dialog.showMessageBox({
+			title: 'Clear thumbnail cache',
+			message: `Deleted ${files.length} thumbnail(s).`
+		});
 	}
 }